genesis-engine/requirements-minimal.txt
# Minimal requirements for API + Dashboard
# Use this for quick start without full AI capabilities

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Core
pydantic>=2.10.0
httpx>=0.27.0

# CLI
rich>=13.0.0

# Environment
python-dotenv>=1.0.0
